perf(search): Add tabs + reset results on new search
